requirements.txt
celery
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
redis
drf_yasg
gunicorn
pytest-django
setuptools
drf-extensions
django-redis
django-debug-toolbar